fix slow loading
The whole `modules-dir = 'dependencies'` thing was done originally
because Deno sometimes bugs when loading symlinked files that
are under `node_modules`. (I'm not sure if it is the TypeScript source
or the implicit Deno runtime environment, but it bugs out with (Syntax
Error: Expected '*' things).

Adding this in `.npmrc` had the side-effect of _massively_ slowing down the
Vite dev server. So much so that it would OOM after a few minutes. There was likely
special logic to handle files in `node_modules`. I have a feeling it is due to some
circular dependencies that are symlinked to each other - in any case V8 Snapshots show
that string concatenations took up a majority of the heap before the segfault.

To work around this, use `node_modules` directory and only for `server-deno`, symlink `dependencies` to `node_modules` to trick Deno into normally loading the files.